backend/requirements.txt
flask
python-dotenv
flask-cors
pyjwt
requests-oauthlib
pyyaml